move archive code into own file
